+++ jumi-actors-maven-plugin/src/it/doesnt-compile-unrelated-sources/src/main/java/example/Unrelated.java
@@ -1,4 +1,0 @@
-package example;
-
-public interface Unrelated {
-}